+++ pkg/encoding/serialize_yaml.go
@@ -9,13 +9,9 @@ func SerializeYAML(pData interface{}) []byte {
 	if err != nil {
 		return nil
 	}
-
 	return res
 }
 
 func DeserializeYAML(pData []byte, pRes interface{}) error {
-	if err := yaml.Unmarshal(pData, pRes); err != nil {
-		return err
-	}
-	return nil
+	return yaml.Unmarshal(pData, pRes)
 }