+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Update `pkg/types`: rename ICommand -> IApp
 - Update `pkg/client/message`: append new checks into IsValid method
 - Update `cmd/hidden_lake/traffic`: append to config settings "key_size_bits" param
+- Update `pkg/encoding`: delete pretty serialize json
 
 <!-- ... -->